// apps/backends/postgres/src/prisma/contract.d.ts
export type CodecTypes = PgTypes;
// ...
export namespace Models {
  export type public_User = {
    id: CodecTypes['pg/int4@1']['output'];
    email: CodecTypes['pg/text@1']['output'];
    username: CodecTypes['pg/text@1']['output'] | null;
    name: CodecTypes['pg/text@1']['output'] | null;
    createdAt: CodecTypes['pg/timestamptz-string@1']['output'];
    updatedAt: CodecTypes['pg/timestamptz-string@1']['output'];
    orders: public_Order[];
    posts: public_Post[];
    readonly [RelationKeys]?: 'orders' | 'posts';
  };
  export type public_Post = {
    id: CodecTypes['pg/int4@1']['output'];
    title: CodecTypes['pg/text@1']['output'];
    content: CodecTypes['pg/text@1']['output'] | null;
    authorId: CodecTypes['pg/int4@1']['output'];
    createdAt: CodecTypes['pg/timestamptz-string@1']['output'];
    updatedAt: CodecTypes['pg/timestamptz-string@1']['output'];
    author: public_User;
    readonly [RelationKeys]?: 'author';
  };
  export type public_Order = {
    id: CodecTypes['pg/int4@1']['output'];
    total: CodecTypes['pg/int4@1']['output'];
    userId: CodecTypes['pg/int4@1']['output'];
    createdAt: CodecTypes['pg/timestamptz-string@1']['output'];
    user: public_User;
    readonly [RelationKeys]?: 'user';
  };
  export type public_Product = {
    id: CodecTypes['pg/int4@1']['output'];
    title: CodecTypes['pg/text@1']['output'];
    price: CodecTypes['pg/int4@1']['output'];
    readonly [RelationKeys]?: never;
  };
}